+++ duck/html/components/script.py
@@ -3,7 +3,6 @@ Script HTML Component.
 
 This module defines a reusable `Script` component for embedding JavaScript code within an HTML document.
 """
-
 from duck.html.components import InnerComponent
 from duck.csp import csp_nonce, csp_nonce_flag
 
@@ -48,14 +47,17 @@ class Script(InnerComponent):
     @property
     def properties(self):
         from duck.settings import SETTINGS
+        
+        # Get original props
         props = super().properties
+        
         # Set CSP configuration.
         if SETTINGS['ENABLE_HEADERS_SECURITY_POLICY']:
             current_nonce = props.get("nonce")
             if not current_nonce:
                 self.set_csp_nonce()
         return props
-          
+        
     def set_csp_nonce(self):
         """
         This tries to retrieve current request nonce.
@@ -74,6 +76,7 @@ class Script(InnerComponent):
             
         # Set CSP configuration
         csp_directives = SETTINGS['CSP_TRUSTED_SOURCES']
+        
         if csp_directives and request:
             script_src = set(csp_directives.get("script-src"))
             if csp_nonce_flag in script_src: